success ERC20 of withdraw

+++ src/main/kotlin/com/api/wallet/controller/dto/request/WithdrawERC20Request.kt
@@ -1,9 +1,9 @@
 package com.api.wallet.controller.dto.request
 
 import com.api.wallet.enums.ChainType
-import java.math.BigInteger
+import java.math.BigDecimal
 
 data class WithdrawERC20Request(
     val chainType: ChainType,
-    val amount: BigInteger,
+    val amount: BigDecimal,
 )
